cfg: return first successor if unable to distingish true- from false-branch
Note, this is a work-around that has to be resolved in the future.
It is an incorrect assumption that the first successor is the true/false
branch, and we only do this so that we can progress with other parts of
the control flow recovery algorithm. The reason that edge information is
lost is that multiple paths may be taken during merge.

See for instance, blake2b_stream.dot

   cfg: unable to locate true branch of edges ("G1_I4" -> "G1_I3") and ("G1_I4" -> "G1_I2") based on edge label; expected "true" and "false", got "" and ""